django_rest_framework_code/with_viewsets/requirements.txt
Django==3.1.7
djangorestframework==3.12.4
PyJWT==2.1.0
